set OoyalaPlayerLayoutController in AbstarctHookActivity Fike
